refactor(block): move block devices to `devices/block`
Move `virtio_block` and `vhost_user_block` under
`devices/block`.
Move `block_common.rs` to `block/mod.rs`.
