+++ src/vmm/src/device_manager/persist.rs
@@ -17,6 +17,15 @@ use super::mmio::*;
 use crate::arch::DeviceType;
 use crate::devices::virtio::balloon::persist::{BalloonConstructorArgs, BalloonState};
 use crate::devices::virtio::balloon::{Balloon, BalloonError};
+use crate::devices::virtio::block::vhost_user::device::VhostUserBlock;
+use crate::devices::virtio::block::vhost_user::persist::{
+    VhostUserBlockConstructorArgs, VhostUserBlockState,
+};
+use crate::devices::virtio::block::vhost_user::VhostUserBlockError;
+use crate::devices::virtio::block::virtio::persist::{
+    VirtioBlockConstructorArgs, VirtioBlockState,
+};
+use crate::devices::virtio::block::virtio::{VirtioBlock, VirtioBlockError};
 use crate::devices::virtio::device::VirtioDevice;
 use crate::devices::virtio::mmio::MmioTransport;
 use crate::devices::virtio::net::persist::{
@@ -28,13 +37,6 @@ use crate::devices::virtio::rng::persist::{
     EntropyConstructorArgs, EntropyPersistError as EntropyError, EntropyState,
 };
 use crate::devices::virtio::rng::Entropy;
-use crate::devices::virtio::vhost_user_block::device::VhostUserBlock;
-use crate::devices::virtio::vhost_user_block::persist::{
-    VhostUserBlockConstructorArgs, VhostUserBlockState,
-};
-use crate::devices::virtio::vhost_user_block::VhostUserBlockError;
-use crate::devices::virtio::virtio_block::persist::{VirtioBlockConstructorArgs, VirtioBlockState};
-use crate::devices::virtio::virtio_block::{VirtioBlock, VirtioBlockError};
 use crate::devices::virtio::vsock::persist::{
     VsockConstructorArgs, VsockState, VsockUdsConstructorArgs,
 };
@@ -650,7 +652,7 @@ mod tests {
 
     use super::*;
     use crate::builder::tests::*;
-    use crate::devices::virtio::block_common::CacheType;
+    use crate::devices::virtio::block::CacheType;
     use crate::resources::VmmConfig;
     use crate::snapshot::Snapshot;
     use crate::vmm_config::balloon::BalloonDeviceConfig;
