+++ src/vmm/src/devices/virtio/block/virtio/persist.rs
@@ -11,11 +11,11 @@ use utils::eventfd::EventFd;
 
 use super::device::DiskProperties;
 use super::*;
+use crate::devices::virtio::block::virtio::device::FileEngineType;
+use crate::devices::virtio::block::virtio::metrics::BlockMetricsPerDevice;
 use crate::devices::virtio::device::{DeviceState, IrqTrigger};
 use crate::devices::virtio::gen::virtio_blk::VIRTIO_BLK_F_RO;
 use crate::devices::virtio::persist::VirtioDeviceState;
-use crate::devices::virtio::virtio_block::device::FileEngineType;
-use crate::devices::virtio::virtio_block::metrics::BlockMetricsPerDevice;
 use crate::devices::virtio::TYPE_BLOCK;
 use crate::logger::warn;
 use crate::rate_limiter::persist::RateLimiterState;
@@ -176,9 +176,9 @@ mod tests {
     use utils::tempfile::TempFile;
 
     use super::*;
+    use crate::devices::virtio::block::virtio::device::VirtioBlockConfig;
     use crate::devices::virtio::device::VirtioDevice;
     use crate::devices::virtio::test_utils::default_mem;
-    use crate::devices::virtio::virtio_block::device::VirtioBlockConfig;
     use crate::snapshot::Snapshot;
 
     #[test]
