+++ src/vmm/src/logger/metrics.rs
@@ -72,10 +72,10 @@ use serde::{Serialize, Serializer};
 use super::FcLineWriter;
 use crate::devices::legacy;
 use crate::devices::virtio::balloon::metrics as balloon_metrics;
+use crate::devices::virtio::block::virtio::metrics as block_metrics;
 use crate::devices::virtio::net::metrics as net_metrics;
 use crate::devices::virtio::rng::metrics as entropy_metrics;
 use crate::devices::virtio::vhost_user_metrics;
-use crate::devices::virtio::virtio_block::metrics as block_metrics;
 use crate::devices::virtio::vsock::metrics as vsock_metrics;
 
 /// Static instance used for handling metrics.
